requirements.txt
fastapi
uvicorn
jinja2
httpx
aiosqlite
python-multipart
pytz
python-dateutil 